Implemented combination of timestamps and rotation in backup file names

+++ plugin/src/main/java/de/intranda/goobi/plugins/RulesetFileUtils.java
@@ -5,6 +5,9 @@ import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.List;
 
 import org.apache.commons.configuration.XMLConfiguration;
 import org.apache.commons.io.FileUtils;
@@ -51,35 +54,67 @@ public abstract class RulesetFileUtils {
     public static void createBackupFile(String fileName) {
         StorageProviderInterface storage = StorageProvider.getInstance();
         try {
-            // Delete oldest file when existing...
-            String lastFileName = RulesetFileUtils.getBackupFileName(fileName, RulesetFileUtils.numberOfBackupFiles);
-            Path lastFile = Paths.get(lastFileName);
-            if (storage.isFileExists(lastFile)) {
-                storage.deleteFile(lastFile);
+            // Delete oldest file if it exists...
+            String lastFileName = RulesetFileUtils.createBackupFileNameWithoutTimestamp(fileName, RulesetFileUtils.numberOfBackupFiles);
+            lastFileName = RulesetFileUtils.backupDirectory + RulesetFileUtils.findBackupFileNameByEnding(lastFileName);
+            if (lastFileName != null) {
+                Path lastFile = Paths.get(lastFileName);
+                if (storage.isFileExists(lastFile)) {
+                    storage.deleteFile(lastFile);
+                }
             }
             // Rename all other backup files...
             // This is the number of the file that should be renamed to the file with the higher number
             int backupId = RulesetFileUtils.numberOfBackupFiles - 1;
             while (backupId > 0) {
-                String newerFileName = RulesetFileUtils.getBackupFileName(fileName, backupId);
-                String olderFileName = RulesetFileUtils.getBackupFileName(fileName, backupId + 1);
-                Path newerFile = Paths.get(newerFileName);
-                if (storage.isFileExists(newerFile)) {
-                    storage.renameTo(newerFile, olderFileName);
+                String newerFileName = RulesetFileUtils.createBackupFileNameWithoutTimestamp(fileName, backupId);
+                newerFileName = RulesetFileUtils.findBackupFileNameByEnding(newerFileName);
+                String newerFileNameWithPath = RulesetFileUtils.backupDirectory + newerFileName;
+                String olderFileName = RulesetFileUtils.createBackupFileNameWithoutTimestamp(fileName, backupId + 1);
+                if (newerFileName != null) {
+                    String timestamp = RulesetFileUtils.extractTimestampFromFileName(newerFileName);
+                    olderFileName = timestamp + "_" + olderFileName;
+                    Path newerFile = Paths.get(newerFileNameWithPath);
+                    if (storage.isFileExists(newerFile)) {
+                        storage.renameTo(newerFile, olderFileName);
+                    }
                 }
                 backupId--;
             }
             // Create backup file...
             String content = RulesetFileUtils.readFile(RulesetFileUtils.rulesetDirectory + fileName);
-            RulesetFileUtils.writeFile(RulesetFileUtils.getBackupFileName(fileName, 1), content);
+            RulesetFileUtils.writeFile(RulesetFileUtils.createBackupFileNameWithTimestamp(fileName, 1), content);
             log.info("Wrote backup file: " + fileName);
         } catch (IOException ioException) {
             log.error(ioException);
         }
     }
 
-    private static String getBackupFileName(String fileName, int backupId) {
-        return RulesetFileUtils.backupDirectory + fileName + "." + backupId;
+    private static String extractTimestampFromFileName(String fileName) {
+        // The date format must be "yyyy_MM_dd_HH_mm_ss"
+        return fileName.substring(0, 19);
+    }
+
+    private static String createBackupFileNameWithoutTimestamp(String fileName, int backupId) {
+        return fileName + "." + backupId;
+    }
+
+    private static String createBackupFileNameWithTimestamp(String fileName, int backupId) {
+        String name = RulesetFileUtils.createBackupFileNameWithoutTimestamp(fileName, backupId);
+        SimpleDateFormat formatter = new SimpleDateFormat("yyyy_MM_dd_HH_mm_ss");
+        String timestamp = formatter.format(new Date());
+        return RulesetFileUtils.backupDirectory + timestamp + "_" + name;
+    }
+
+    private static String findBackupFileNameByEnding(String ending) {
+        StorageProviderInterface storage = StorageProvider.getInstance();
+        List<String> fileNames = storage.list(RulesetFileUtils.backupDirectory);
+        for (int index = 0; index < fileNames.size(); index++) {
+            if (fileNames.get(index).endsWith(ending)) {
+                return fileNames.get(index);
+            }
+        }
+        return null;
     }
 
     public static String readFile(String fileName) {
